Added GL4 configurations to autobuilder
Binaries submitted in 1040471
Build time : 11 minutes 8.807000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/": change = 1040464]

+++ Tr2ShaderALGL4.cpp
@@ -44,7 +44,7 @@ static int CreateShader( Tr2RenderContextEnum::ShaderType type, const void* byte
 	ON_BLOCK_EXIT( [&]{ if( shader ) { glDeleteShader( shader ); } } );
 
 	static const char* prefix = "#version 410 core\n";
-	static const GLint prefixLength = strlen( prefix );
+	static const GLint prefixLength = GLint( strlen( prefix ) );
 
 	const char* code[] = { prefix, reinterpret_cast<const char*>( bytecode ) };
 	GLint length[] = { prefixLength, GLint( bytecodeSize ) };
